requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
alembic
psycopg2-binary
slowapi
python-dotenv>=1.0.0
passlib
python-multipart>=0.0.6
email-validator
faiss-cpu
numpy>=1.21.0
gradio>=4.0.0
pydantic>=2.0.0
sentence-transformers>=2.2.2
Pillow>=10.0.0
openai>=1.0.0
requests>=2.31.0
pymupdf>=1.23.0
fitz
pdf2image>=1.16.0
poppler-utils
